# tutorial/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_evidence_api_v1_incidents__incident_id__evidence_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_evidence_api_v1_incidents__incident_id__evidence_post",
        "type": "object"
      },
...
      "EvidenceResponse": {
        "description": "Evidence metadata.",
        "properties": {
          "collected_by": {
            "title": "Collected By",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "custody_chain": {
            "items": {},
            "title": "Custody Chain",
            "type": "array"
          },
          "evidence_type": {
            "title": "Evidence Type",
            "type": "string"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "file_size_bytes": {
            "title": "File Size Bytes",
            "type": "integer"
          },
          "hash_sha256": {
            "title": "Hash Sha256",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "storage_location": {
            "title": "Storage Location",
            "type": "string"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "id",
          "incident_id",
          "evidence_type",
          "file_path",
          "hash_sha256",
          "file_size_bytes",
          "metadata",
          "collected_by",
          "custody_chain",
          "storage_location",
          "created_at",
          "verified_at"
        ],
        "title": "EvidenceResponse",
        "type": "object"
      },
      "EvidenceType": {
        "enum": [
          "file",
          "memory",
          "network",
          "disk",
          "log",
          "other"
        ],
        "title": "EvidenceType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncidentCreate": {
        "additionalProperties": false,
        "description": "Payload for creating a persisted incident.",
        "properties": {
          "assigned_agents": {
            "items": {},
            "title": "Assigned Agents",
            "type": "array"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "incident_type": {
            "default": "unknown",
            "maxLength": 128,
            "title": "Incident Type",
            "type": "string"
          },
          "raw_evidence_refs": {
            "items": {},
            "title": "Raw Evidence Refs",
            "type": "array"
          },
          "severity": {
            "$ref": "#/components/schemas/IncidentSeverity"
          },
          "source_ip": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ip"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus",
            "default": "open"
          },
          "tags": {
            "items": {},
            "title": "Tags",
            "type": "array"
          },
          "target_asset": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Asset"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "severity"
        ],
        "title": "IncidentCreate",
        "type": "object"
      },
      "IncidentDetail": {
        "description": "Incident with nested investigation and evidence context.",
        "properties": {
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "incident": {
            "$ref": "#/components/schemas/IncidentResponse"
          },
          "investigation_steps": {
            "items": {
              "$ref": "#/components/schemas/InvestigationStepResponse"
            },
            "title": "Investigation Steps",
            "type": "array"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonSummaryResponse"
            },
            "title": "Lessons",
            "type": "array"
          }
        },
        "required": [
          "incident",
          "investigation_steps",
          "evidence",
          "lessons"
        ],
        "title": "IncidentDetail",
        "type": "object"
      },
      "IncidentResponse": {
        "additionalProperties": false,
        "description": "Incident summary returned from the API.",
        "properties": {
          "assigned_agents": {
            "items": {},
            "title": "Assigned Agents",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_type": {
            "title": "Incident Type",
            "type": "string"
          },
          "raw_evidence_refs": {
            "items": {},
            "title": "Raw Evidence Refs",
            "type": "array"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ip"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {},
            "title": "Tags",
            "type": "array"
          },
          "target_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Asset"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "severity",
          "status",
          "source_ip",
          "target_asset",
          "incident_type",
          "raw_evidence_refs",
          "created_at",
          "updated_at",
          "completed_at",
          "assigned_agents",
          "tags"
        ],
        "title": "IncidentResponse",
        "type": "object"
      },
      "IncidentSeverity": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "IncidentSeverity",
        "type": "string"
      },
      "IncidentStatus": {
        "enum": [
          "open",
          "triaging",
          "investigating",
          "contained",
          "resolved",
          "closed"
        ],
        "title": "IncidentStatus",
        "type": "string"
      },
...
      "InvestigationStepResponse": {
        "additionalProperties": false,
        "description": "Single investigation step.",
        "properties": {
          "action_taken": {
            "title": "Action Taken",
            "type": "string"
          },
          "agent_name": {
            "title": "Agent Name",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "correction_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction Reason"
          },
          "execution_time_ms": {
            "title": "Execution Time Ms",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "interpretation": {
            "title": "Interpretation",
            "type": "string"
          },
          "is_self_correction": {
            "title": "Is Self Correction",
            "type": "boolean"
          },
          "raw_output": {
            "title": "Raw Output",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "tool_used": {
            "title": "Tool Used",
            "type": "string"
          }
        },
        "required": [
          "id",
          "incident_id",
          "agent_name",
          "action_taken",
          "tool_used",
          "raw_output",
          "interpretation",
          "confidence",
          "timestamp",
          "is_self_correction",
          "correction_reason",
          "execution_time_ms"
        ],
        "title": "InvestigationStepResponse",
        "type": "object"
      },
...
      "LessonSummaryResponse": {
        "additionalProperties": false,
        "description": "Minimal lesson row.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "difficulty": {
            "title": "Difficulty",
            "type": "string"
          },
          "estimated_duration_minutes": {
            "title": "Estimated Duration Minutes",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "incident_id",
          "title",
          "difficulty",
          "estimated_duration_minutes",
          "created_at"
        ],
        "title": "LessonSummaryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The World's First Agentic AI That Learns by Teaching",
    "title": "TUTORIAL API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/incidents/": {
      "get": {
        "description": "List incidents with optional filters.",
        "operationId": "list_incidents_api_v1_incidents__get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  },
                  "title": "Response List Incidents Api V1 Incidents  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Incidents",
        "tags": [
          "incidents"
        ]
      },
      "post": {
        "description": "Submit a new security incident for investigation.",
        "operationId": "create_incident_api_v1_incidents__post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Incident",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "description": "Return incident detail including investigation and evidence.",
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/accuracy-report": {
      "get": {
        "description": "Return stored FIND EVIL! accuracy metadata.",
        "operationId": "get_accuracy_report_api_v1_incidents__incident_id__accuracy_report_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Accuracy Report Api V1 Incidents  Incident Id  Accuracy Report Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Accuracy Report",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/evidence": {
      "post": {
        "description": "Persist an uploaded evidence file and metadata.",
        "operationId": "upload_evidence_api_v1_incidents__incident_id__evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "evidence_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EvidenceType",
              "default": "file"
            }
          },
          {
            "in": "query",
            "name": "collected_by",
            "required": false,
            "schema": {
              "default": "api-client",
              "title": "Collected By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "storage_location",
            "required": false,
            "schema": {
              "default": "local",
              "title": "Storage Location",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_evidence_api_v1_incidents__incident_id__evidence_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Upload Evidence Api V1 Incidents  Incident Id  Evidence Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Evidence",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}/investigate": {
      "post": {
        "description": "Run the autonomous defense workflow for a persisted incident.",
        "operationId": "start_investigation_api_v1_incidents__incident_id__investigate_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Start Investigation Api V1 Incidents  Incident Id  Investigate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Investigation",
        "tags": [
          "incidents"
        ]
      }
    },
...
  }
}